ammend ns config
